add stdint

+++ workshop/hardware/expressions-and-control-flow/workshop/CodeBlocks/20.c
@@ -1,7 +1,8 @@
 #include <stdio.h>
+#include <stdint.h>
 
 int main() {
-	int z = 13;
+	uint8_t z = 13;
 	// if z is between 10 and 20 print 'Sweet!'
 	// if less than 10 print 'More!',
 	// if more than 20 print 'Less!'
